Heavy rework
Adjusted much of the core logic in how this mod works and added way toooo many options.
Fixed the blackout shader and replaced it with Flux to allow for more customizability.
Separated out the thresholds for entering and leaving Drowsy and Sleep.
Added more features for detecting when you are awake. (Head or Player Movement, Controller Movement, Menus opening)
Added a Debounce to make it harder to accidently transition to the awake state.
Added the option to lower volume when sleeping
Added integration with avatar Parameters (BlackoutModDrowsy & BlackoutModSleep)